+++ cpp/src/text/normalize.cu
@@ -184,7 +184,7 @@ std::unique_ptr<cudf::column> normalize_spaces(
   auto d_chars = chars_column->mutable_view().data<char>();
 
   // copy tokens to the chars buffer
-  thrust::for_each_n(rmm::exec_policy(stream)->on(stream.value()),
+  thrust::for_each_n(rmm::exec_policy(stream),
                      thrust::make_counting_iterator<cudf::size_type>(0),
                      strings_count,
                      normalize_spaces_fn{d_strings, d_offsets, d_chars});
@@ -247,7 +247,7 @@ std::unique_ptr<cudf::column> normalize_characters(cudf::strings_column_view con
 
   // build the chars output data: convert the 4-byte code-point values into UTF-8 chars
   thrust::for_each_n(
-    rmm::exec_policy(stream)->on(stream.value()),
+    rmm::exec_policy(stream),
     thrust::make_counting_iterator<cudf::size_type>(0),
     strings_count,
     codepoint_to_utf8_fn{*strings_column, cp_chars, cp_offsets, d_offsets, d_chars});